# docs/swagger.yaml
basePath: /
definitions:
  apool.APool:
    properties:
      nextNum:
        type: integer
      numToAttrib:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
    type: object
...
  errors.Error:
    description: Standardized API error response
    properties:
      error:
        example: 404
        type: integer
      message:
        example: Pad not found
        type: string
    type: object
  groups.CreateGroupPadRequest:
    properties:
      authorId:
        type: string
      padName:
        type: string
      text:
        type: string
    type: object
  groups.GroupIDResponse:
    properties:
      groupID:
        type: string
    type: object
...
  pad.AttributePoolResponse:
    properties:
      pool:
        $ref: '#/definitions/apool.APool'
    type: object
...
  pad.SetTextRequest:
    properties:
      authorId:
        type: string
      text:
        type: string
    type: object
  pad.TextResponse:
    properties:
      text:
        type: string
    type: object
...
host: localhost:9001
info:
  contact:
    email: support@etherpad.org
    name: Etherpad Support
    url: https://github.com/ether/etherpad-go
  description: REST API for Etherpad Go - Collaborative Text Editor
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Etherpad Go API
  version: "1.0"
paths:
...
  /admin/api/groups:
    post:
      consumes:
      - application/json
      description: Creates a new group and returns its ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/groups.GroupIDResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Create a new group
      tags:
      - Groups
  /admin/api/groups/{groupId}:
    delete:
      consumes:
      - application/json
      description: Deletes a group
      parameters:
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Delete a group
      tags:
      - Groups
  /admin/api/groups/{groupId}/pads:
    post:
      consumes:
      - application/json
      description: Creates a new pad within a group
      parameters:
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: string
      - description: Pad data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/groups.CreateGroupPadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Create a pad in a group
      tags:
      - Groups
...
  /admin/api/pads/{padId}/{rev}/revisionChangeset:
    get:
      consumes:
      - application/json
      description: Returns the changeset of a specific revision of a pad
      parameters:
      - description: Pad ID
        in: path
        name: padId
        required: true
        type: string
      - description: Revision number
        in: path
        name: rev
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Get revision changeset
      tags:
      - Pads
...
  /admin/api/pads/{padId}/attributePool:
    get:
      consumes:
      - application/json
      description: Returns the attribute pool of a pad
      parameters:
      - description: Pad ID
        in: path
        name: padId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pad.AttributePoolResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Get attribute pool of a pad
      tags:
      - Pads
...
  /admin/api/pads/{padId}/text:
    get:
      consumes:
      - application/json
      description: Returns the current text of a pad, optionally for a specific revision
      parameters:
      - description: Pad ID
        in: path
        name: padId
        required: true
        type: string
      - description: Revision number
        in: query
        name: rev
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pad.TextResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Get pad text
      tags:
      - Pads
    post:
      consumes:
      - application/json
      description: Updates the text of a pad
      parameters:
      - description: Pad ID
        in: path
        name: padId
        required: true
        type: string
      - description: Text and Author ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pad.SetTextRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Set pad text
      tags:
      - Pads
...
  /p/{pad}/{rev}/export/{type}:
    get:
      description: Exports the content of a pad to various formats (pdf, word, txt,
        html, open, etherpad, markdown)
      parameters:
      - description: Pad ID
        in: path
        name: pad
        required: true
        type: string
      - description: Revision number
        in: path
        name: rev
        type: string
      - description: Export type (pdf, word, txt, html, open, etherpad, markdown)
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Exported file
          schema:
            type: file
        "400":
          description: Invalid export type
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Pad not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
        "503":
          description: Export not available
          schema:
            type: string
      summary: Export a pad
      tags:
      - Export
  /p/{pad}/export/{type}:
    get:
      description: Exports the content of a pad to various formats (pdf, word, txt,
        html, open, etherpad, markdown)
      parameters:
      - description: Pad ID
        in: path
        name: pad
        required: true
        type: string
      - description: Export type (pdf, word, txt, html, open, etherpad, markdown)
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Exported file
          schema:
            type: file
        "400":
          description: Invalid export type
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Pad not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
        "503":
          description: Export not available
          schema:
            type: string
      summary: Export a pad
      tags:
      - Export
...
securityDefinitions:
  BearerAuth:
    description: Bearer Token for Admin API authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"